+++ server/src/tangible/models/__init__.py
@@ -30,7 +30,13 @@ from tangible.models.notification import NotificationPreference
 from tangible.models.photo import Photo
 from tangible.models.scraper_registry_pin import ScraperRegistryPin
 from tangible.models.share_link import ShareLink
-from tangible.models.shopping import ListType, ShoppingItem, ShoppingStore, ShoppingStoreAisle
+from tangible.models.shopping import (
+    BarcodeHint,
+    ListType,
+    ShoppingItem,
+    ShoppingStore,
+    ShoppingStoreAisle,
+)
 from tangible.models.sync_doc import AutomergeChange, AutomergeDoc
 from tangible.models.tag import ItemTag, Tag
 from tangible.models.user import (
@@ -48,6 +54,7 @@ __all__ = [
     "AuditLogEntry",
     "AutomergeChange",
     "AutomergeDoc",
+    "BarcodeHint",
     "BundleAsset",
     "BundleItem",
     "Category",